backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
pillow
tensorflow
numpy
exifread
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
